+++ fixtures/ballot.json
@@ -1,5 +1,5 @@
 {
-   "ballot_hash":"f2f11a3ad2c03e7dea6d47316e607e4807a18ae064ddf6195fb27a0a180e9168dd9f0f169a352f409e32ee3f89b41a9a21a618a4857788efba13382066aa8df6",
+   "ballot_hash":"e182b0ee3654d252fec6b47b210e8e2a5ce051b6702ae1da6516dab6a47f6df507b0ce6dfd0d9b267c9e5ede8be8732b365365f4e8cb39159d75d7f4a9b80cc2",
    "replication": {
       "issue_date":"7/10/2021",
       "choices":[
@@ -175,7 +175,7 @@
               },
               {
                  "key_type": "P2048",
-                 "public_key":"49255557909566377757205787942241920039300614692228631933601582821106871068005356966433637214118407121245354594153824991276653304636628584850713636052557060219166337409263662901005610924779652768416782653063104603596145563915869808761324667445814546620159061672864086092992572276054930288759193130207398313556889441768712811418707815590377969200124994502555199142105124140414055104449381909211225399423619114132430929198004305153560570240084213683322189831030472103223092473610276457748166677617242297272670929119424923393774252900481085781723237077938740536050524204345920190173074197841944187454308298978583594373295",
+                 "public_key":"49255557909566377757205787942241920039300614692228631933601582821106871068005356966433637214118407121245354594153824991276653304636628584850713636052557060219166337409263662901005610924779652768416782653063104603596145563915869808761324667445814546620159061672864086092992572276054930288759193130207398313556889441768712811418707815590377969200124994502555199142105124140414055104449381909211225399423619114132430929198004305153560570240084213683322189831030472103223092473610276457748166677617242297272670929119424923393774252900481085781723237077938740536050524204345920190173074197841944187454308298978583594373295"
               }
            ]
         }
